chore : 메인/마이 페이지 커스텀훅 작업, 자잘한 CSS추가 및 수정
