+++ src/components/mypage/PagingTap.jsx
@@ -10,7 +10,7 @@ function PagingTap({data, btnInfo, currentBtn, buttonClickHandler}) {
 
   return (
     <Div width="100%" marginTop="3rem" gap="3rem">
-        <Div fDirection="row" width="100%" jc="space-around" borderBottom="1px solid #e6e6e6">
+        <Div fDirection="row" width="100%" borderBottom="1px solid #e6e6e6" gap="5rem">
             {btnInfo.map((item) => 
                 <LPBtn
                 key={nanoid()}
@@ -44,6 +44,7 @@ function PagingTap({data, btnInfo, currentBtn, buttonClickHandler}) {
 export default PagingTap
 
 const LPBtn = styled.button`
+    width: 100%;
     border: none;
     font-size: 1rem;
     padding-bottom: 8px;
