+++ src/hooks/useGetMainPageData.js
@@ -1,0 +1,28 @@
+import { useQuery } from "@tanstack/react-query";
+import axios from "axios";
+import { getCookie } from "../shared/Cookies";
+
+const useGetMainPageData = () => {
+    const { data } = useQuery({
+        queryKey: ["GET_MAINPAGE"],
+        queryFn: async () => {
+          const token = getCookie("token");
+          if(!token){
+            const res = await axios.get(`${process.env.REACT_APP_SERVER_URL}/main`)
+            return res.data.data;
+          } else {
+            const res = await axios.get(`${process.env.REACT_APP_SERVER_URL}/main`, {
+              headers: {
+                Authorization: `Bearer ${token}`,
+              }
+            });
+            return res?.data.data;
+          }
+        }
+    })
+    return {
+      data
+    }
+}
+
+export default useGetMainPageData;
